+++ test/test_error.py
@@ -5,16 +5,17 @@ from aio_overpass import Client, Query
 from aio_overpass.error import (
     CallError,
     GiveupError,
-    QueryError,
     QueryLanguageError,
     QueryRejectCause,
     QueryRejectError,
+    QueryResponseError,
     ResponseError,
+    RunnerError,
 )
-from aio_overpass.query import DefaultQueryRunner
+from aio_overpass.query import DefaultQueryRunner, QueryRunner
 
 import pytest
-from aioresponses import CallbackResult, aioresponses
+from aioresponses import aioresponses
 
 
 URL_INTERPRETER = re.compile(r"^https://overpass-api\.de/api/interpreter\?data=.+$")
@@ -37,15 +38,16 @@ def mock_response():
             url=URL_STATUS,
             body=mock_status,
             status=200,
+            repeat=True,
         )
 
         yield m
 
 
 @pytest.mark.asyncio
-async def mock_run_query(mock_response, body, content_type):
+async def mock_run_query(mock_response, body, content_type, **kwargs):
     c = Client(runner=DefaultQueryRunner(max_tries=1))
-    q = Query("")
+    q = Query("", **kwargs)
 
     mock_response.get(
         url=URL_INTERPRETER,
@@ -87,10 +89,10 @@ async def test_too_many_queries(mock_response):
     expected = [
         "runtime error: open64: 0 Success /osm3s_v0.7.54_osm_base Dispatcher_Client::request_read_and_idx::rate_limited. Please check /api/status for the quota of your IP address."
     ]
-    assert err.value.messages == expected
+    assert err.value.remarks == expected
 
-    _ = str(err)
-    _ = repr(err)
+    _ = str(err.value)
+    _ = repr(err.value)
 
 
 @pytest.mark.asyncio
@@ -122,10 +124,10 @@ async def test_too_busy(mock_response):
     expected = [
         "runtime error: open64: 0 Success /osm3s_v0.7.54_osm_base     Dispatcher_Client::request_read_and_idx::timeout. The server is probably too busy to handle your request."
     ]
-    assert err.value.messages == expected
+    assert err.value.remarks == expected
 
-    _ = str(err)
-    _ = repr(err)
+    _ = str(err.value)
+    _ = repr(err.value)
 
 
 @pytest.mark.asyncio
@@ -148,16 +150,16 @@ async def test_other_query_error(mock_response):
 </html>
     """
 
-    with pytest.raises(QueryError) as err:
-        await mock_run_query(mock_response, body, content_type="text/html")
+    with pytest.raises(QueryResponseError) as err:
+        await mock_run_query(mock_response, body, content_type="text/html", my_kwarg=42)
 
     expected = [
         "runtime error: open64: 2 No such file or directory /osm3s_v0.7.54_osm_base Dispatcher_Client::1"
     ]
-    assert err.value.messages == expected
+    assert err.value.remarks == expected
 
-    _ = str(err)
-    _ = repr(err)
+    _ = str(err.value)
+    _ = repr(err.value)
 
 
 @pytest.mark.asyncio
@@ -211,10 +213,10 @@ async def test_syntax_error(mock_response):
         r"""line 1: parse error: ',' or ']' expected - '%' found.""",
         r"""line 1: static error: For the attribute "k" of the element "has-kv" the only allowed values are non-empty strings.""",
     ]
-    assert err.value.messages == expected
+    assert err.value.remarks == expected
 
-    _ = str(err)
-    _ = repr(err)
+    _ = str(err.value)
+    _ = repr(err.value)
 
 
 @pytest.mark.asyncio
@@ -237,16 +239,16 @@ async def test_other_query_error_remark(mock_response):
 }
     """
 
-    with pytest.raises(QueryError) as err:
+    with pytest.raises(QueryResponseError) as err:
         await mock_run_query(mock_response, body, content_type="application/json")
 
     expected = [
         "runtime error: Way 547230203 cannot be expanded at timestamp 2018-05-08T15:48:01Z."
     ]
-    assert err.value.messages == expected
+    assert err.value.remarks == expected
 
-    _ = str(err)
-    _ = repr(err)
+    _ = str(err.value)
+    _ = repr(err.value)
 
 
 @pytest.mark.asyncio
@@ -273,10 +275,10 @@ async def test_exceeded_maxsize(mock_response):
     expected = [
         r"""runtime error: Query run out of memory in "recurse" at line 1 using about 541 MB of RAM.""",
     ]
-    assert err.value.messages == expected
+    assert err.value.remarks == expected
 
-    _ = str(err)
-    _ = repr(err)
+    _ = str(err.value)
+    _ = repr(err.value)
 
 
 @pytest.mark.asyncio
@@ -303,10 +305,10 @@ async def test_exceeded_timeout(mock_response):
     expected = [
         r"""runtime error: Query timed out in "query" at line 3 after 2 seconds.""",
     ]
-    assert err.value.messages == expected
+    assert err.value.remarks == expected
 
-    _ = str(err)
-    _ = repr(err)
+    _ = str(err.value)
+    _ = repr(err.value)
 
 
 @pytest.mark.asyncio
@@ -317,8 +319,8 @@ async def test_connection_refused():
     with aioresponses(), pytest.raises(CallError) as err:
         await c.run_query(q)
 
-    _ = str(err)
-    _ = repr(err)
+    _ = str(err.value)
+    _ = repr(err.value)
 
 
 @pytest.mark.asyncio
@@ -330,20 +332,150 @@ async def test_internal_server_error():
         m.get(URL_STATUS, status=500, repeat=True)
         await c.run_query(q)
 
-    _ = str(err)
-    _ = repr(err)
+    _ = str(err.value)
+    _ = repr(err.value)
 
 
 @pytest.mark.asyncio
 async def test_timeout_error():
     c = Client(runner=DefaultQueryRunner(max_tries=1))
     q = Query("")
+    q2 = Query("", my_kwarg=42)
 
     with aioresponses() as m:
         m.get(URL_STATUS, exception=asyncio.TimeoutError())
-
         with pytest.raises(GiveupError) as err:
             await c.run_query(q)
 
-    _ = str(err)
-    _ = repr(err)
+        m.get(URL_STATUS, exception=asyncio.TimeoutError())
+        with pytest.raises(GiveupError) as err2:
+            await c.run_query(q2)
+
+    _ = str(err.value)
+    _ = repr(err.value)
+    _ = str(err2.value)
+    _ = repr(err2.value)
+
+
+@pytest.mark.asyncio
+async def test_runner_error():
+    class BadQueryRunner(QueryRunner):
+        async def __call__(self, query: Query) -> None:
+            raise RuntimeError
+
+    c = Client(runner=BadQueryRunner())
+    q = Query("")
+
+    with pytest.raises(RunnerError) as err:
+        await c.run_query(q)
+
+    assert isinstance(err.value.cause, RuntimeError)
+
+    _ = str(err.value)
+    _ = repr(err.value)
+
+
+@pytest.mark.asyncio
+async def test_status_error(mock_response):
+    c = Client(runner=DefaultQueryRunner(max_tries=1))
+    q = Query("")
+
+    mock_response.get(
+        url=URL_INTERPRETER,
+        body="{}",
+        status=429,
+        content_type="application/json",
+    )
+    with pytest.raises(QueryRejectError) as err1:
+        await c.run_query(q)
+
+    assert err1.value.cause is QueryRejectCause.TOO_MANY_QUERIES
+
+    mock_response.get(
+        url=URL_INTERPRETER,
+        body="{}",
+        status=504,
+        content_type="application/json",
+    )
+    with pytest.raises(QueryRejectError) as err2:
+        await c.run_query(q)
+
+    assert err2.value.cause is QueryRejectCause.TOO_BUSY
+
+    _ = str(err1.value)
+    _ = repr(err1.value)
+    _ = str(err2.value)
+    _ = repr(err2.value)
+
+
+@pytest.mark.asyncio
+async def test_unexpected_message_error(mock_response):
+    c = Client(runner=DefaultQueryRunner(max_tries=1))
+    q = Query("")
+
+    msg = "something that does not match a ql error"
+    body = rf"""
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN"
+    "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en">
+<head>
+  <meta http-equiv="content-type" content="text/html; charset=utf-8" lang="en"/>
+  <title>OSM3S Response</title>
+</head>
+<body>
+
+<p>The data included in this document is from www.openstreetmap.org. The data is made available under ODbL.</p>
+<p><strong style="color:#FF0000">Error</strong>: {msg} </p>
+
+</body>
+</html>
+    """
+
+    mock_response.get(
+        url=URL_INTERPRETER,
+        body=body,
+        status=400,
+        content_type="text/html",
+    )
+    with pytest.raises(QueryResponseError) as err:
+        await c.run_query(q)
+
+    _ = str(err.value)
+    _ = repr(err.value)
+
+
+@pytest.mark.asyncio
+async def test_no_message_error(mock_response):
+    c = Client(runner=DefaultQueryRunner(max_tries=1))
+    q = Query("")
+
+    msg = "something that does not match a ql error"
+    body = rf"""
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN"
+    "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en">
+<head>
+  <meta http-equiv="content-type" content="text/html; charset=utf-8" lang="en"/>
+  <title>OSM3S Response</title>
+</head>
+<body>
+
+<p>The data included in this document is from www.openstreetmap.org. The data is made available under ODbL.</p>
+
+</body>
+</html>
+    """
+
+    mock_response.get(
+        url=URL_INTERPRETER,
+        body=body,
+        status=200,
+        content_type="text/html",
+    )
+    with pytest.raises(ResponseError) as err:
+        await c.run_query(q)
+
+    _ = str(err.value)
+    _ = repr(err.value)